feat: Implement turbofish operator (#3542)
# Description

## Problem\*

Resolves #3413 

## Summary\*

Adds the turbofish operator `::<>` to variables and method calls in
Noir.

## Additional Context

I'm publishing this as a draft because I'm shelving this work for now
after implementing the majority of it.

The work that remains to be done is handling the special case of method
calls on generic impls. A generic impl will implicitly add the impl
generics to each method - which means the generic count on a method
itself as seen by the turbofish operator will not match the actual
generics on the function internally. We'll likely need to separate out
these implicit generics internally. Such that `expected_generic_count =
function_generics - impl_generics`. I've added this as a test case to
the `generics` test to ensure it works when this is merged.

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[Exceptional Case]** Documentation to be submitted in a separate
PR.
- [x] No documentation is in this PR yet since it is still a draft. 

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>
Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>