chore(experimental): Elaborate globals (#5069)
# Description

## Problem\*

Working towards #4992 

## Summary\*

Adds support for elaborating globals. This was the last major item for
resolution + type checking in the elaborator.

## Additional Context

Next up in the elaborator will be integrating the scanning
pass/interpreter and hooking it up to our integration tests.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.