fix: Use plain integer addresses for opcodes in DAP disassembly view (#4941)
# Description

## Problem\*

Resolves #4904

The disassembly view from VS.Code didn't work properly. Sometimes it
would only show the first opcode only and once open it wouldn't update
the current opcode pointer when stepping through the code. This was
because we were using opcode locations (eg. `3.17`) as instruction
references (typed as string in the API), but VS.Code expects integers
either in decimal or hexadecimal format.

## Summary\*

This PR updates the disassembly request DAP command handler to return
integer addresses only, and changes the instruction set breakpoint
command to parse and map the given address to the appropriate opcode
location.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.