feat: Activate return_data in ACIR opcodes (#5080)
# Description

## Problem\*

Related to #4974

## Summary\*
This PR mark return data array with return_data block type, and disable
the unused arrays optimisation for databus.
This is required because the return_data is not returned from main
although similar constraints will be generated on the backend side.


## Additional Context
This should be the last piece for this databus iteration from issue
#4974, and it will allow to generate an end-to-end test with the databus
to check if we get what is expected.


## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.