fix: remove finalize from acir create circuit (#6585)
Largely undoing the changes made in
[this](https://github.com/AztecProtocol/aztec-packages/pull/6164) PR
that added a call to `finalize_circuit` in the create_circuit methods in
acir_format. This change was originally made in order to facilitate
accurate final gate counts (previously handled via a hack). The finalize
model has been reverted and the hack has been replaced with a method
that computes the number of gates added by the `add_gates_to_ensure..`
methods in the builders. This change was necessary in order to allow
gates to be added to a circuit generated from acir, which is needed in
ClientIvc accumulation.