feat: batch simulate (#6599)
Closes: https://github.com/AztecProtocol/aztec-packages/issues/6549

Also removes `FunctionData` from `TxExecutionRequest` and a bunch of
other places, in favor of just the `FunctionSelector`(which is actually
what's needed). This was missed from #3417 and was driving me crazy when
trying to add a few more things to `FunctionCall`. It had also leaked
somehow into the AVM enqueued calls.