feat: ACIR integration tests in Bb (#6607)
This adds a GoogleTest suite for running the ACIR tests generated
through Noir. They currently run with GUH arithmetization but they could
easily be extended.

To handle a linker error (and because WASM compilation required this
through an implicit dependency not compiling) I move some function
definitions from a header to a source file and take the opportunity to
reduce the header size as well.