feat: introduce initialize_or_replace (#6519)
This fixes the issue described in [these
comments](https://github.com/AztecProtocol/aztec-packages/pull/6442#discussion_r1602298155),
which highlighted that `PrivateMutable` was providing a poor API as it
forced users to call unconstrained functions and leave correctness up to
them. This is a simple replacement to make #6442 cleaner and help guide
future development, but it should only be considered a patch - we'll
eventually do a more thoughtful complete pass over the entire API to
bring coherency to it.

I tried adding tests for this (I wanted to test that regardless what the
return value of the nullifier check oracle was, either an inclusion
proof or a new nullifier would be included in the context), but this
proved to be very hard because notes are sent to the oracle as raw
fields in a custom packed format which `oracle::get_notes` decodes on
the spot. We'd need to have that be a separate library, but that exceeds
the scope of this PR.