feat: Consider block parameters in variable liveness (#5097)
# Description

## Problem\*

Previously, on variable liveness we completely ignored block parameters,
by allocating all the parameters of all blocks at the start of the
function codegen and maintaining them allocated for the whole function.

## Summary\*

With this change, we add a definition point for block parameters and we
start considering them and allocating and freeing them dynamically. This
should save space in the stack, especially in heavily inlined functions
with many blocks.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.