chore: perform dead instruction elimination through `std::as_witness` (#5123)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds handling for `std::as_witness` in dead instruction
elimination. i.e. if we're asking for an SSA value to be assigned to a
witness when that value would not otherwise be represented in the ACIR
then we can remove it.

This issue can be seen in #5122 where making two calls to the same
function with the same inputs and asserting that the outputs are the
same results in constraints whereas this can be determined at runtime.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.