feat: make ACVM generic across fields (#5114)
# Description

Step towards #5055 

## Summary\*

This PR starts us down the road towards removing the compile-time flags
for specifying which field is being used by replacing the usage of the
`FieldElement` type with an `AcirField` trait. This trait is mostly all
of the external methods from `FieldElement` dumped into it for now but
we can break it down in future PRs (XOR and AND are looking ready for
culling)

I've taken this route rather than just using `generic_ark::FieldElement`
as we'd need to have trait bounds for `PrimeField` anyway so we might as
well have our own trait.

I've had to delete a couple of trait implementations which fall afoul of
the orphan rule now it's being implemented for a generic type (e.g. we
need to use `MemoryValue::new_field` explicitly now) but nothing too
bad.

## Additional Context


## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.