fix(experimental elaborator): Fix duplicate `resolve_type` on self type and don't leak a trait impl's generics (#5102)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Fixes two issues:
- Use cached `resolved_object_type` instead of re-resolving the self
type in `declare_methods_on_struct` and `check_trait_impl_coherence`.
This is needed to prevent duplicate errors, and in the later case, since
generics are not added to scope to be able to re-resolve the object type
at that point.
- Fix a case when resolving trait impls that `self.generics` was not
cleared afterward.

## Additional Context

Down to 112 errors in the standard library now. The remaining errors are
variants of:
- "Duplicate definitions of \<global> found"
- "Duplicate definitions of \<generic> found" - looks to be only one
case where there is a generic directly on a trait method. `fn
hash<H>(self, state: &mut H) where H: Hasher;`
- "Expression type is ambiguous" - from the two turbofish cases that
haven't been merged in this branch yet
- "Expected type &mut _, found type H" - issues with auto-deref and
trait methods possibly
- "No matching impl found for ..." - trait solver issues

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: Maxim Vezenov <mvezenov@gmail.com>
Co-authored-by: Álvaro Rodríguez <sirasistant@gmail.com>
Co-authored-by: guipublic <47281315+guipublic@users.noreply.github.com>