chore: run all test programs in brillig as well as ACIR (#5128)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR helps ensure consistency between ACIR and brillig by checking
that every `execution_success` test case were compiled and run in
brillig then it would succeed.

This has thrown up a couple of test case failures which I've added to a
list to be ignored for now. We can address these in follow-up PRs

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.