fix(experimental elaborator): Fix global values used in the elaborator (#5135)
# Description

## Problem\*

Resolves an issue where globals would not have their type set correctly
in the elaborator. This is because `elaborate_let` was called which
creates a fresh DefinitionId rather than the one already defined by the
global, and pushes the type to that fresh id.

## Summary\*

Adds a `elaborate_global_let` function which copies over the new type to
the original DefinitionId defined by the global.
This PR also adds a call to `self.type_variables.clear()` after a global
is resolved so that its type is not defaulted too early.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.