chore(ci): don't raise MSRV issue if workflow cancelled (#5143)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We're currently getting a large number of false positive alerts such as
https://github.com/noir-lang/noir/issues/5137 due to the workflow being
cancelled when multiple PRs are merged in quick succession.

This PR changes the workflow to treat cancellation as a success rather
than a failure to avoid this.


## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.