fix(experimental elaborator): Fix definition kind of globals and tuple patterns with `--use-elaborator` flag (#5139)
# Description

## Problem\*

Resolves two issues:
1. A type error on some tuple patterns because a `follow_bindings` call
was missing
2. When elaborating global lets we were setting the existing definition
id to a DefinitionKind::Global, but it already was one. The fresh ID
created by the let should have been marked global instead.

## Summary\*



## Additional Context

After this PR and https://github.com/noir-lang/noir/pull/5138 all
test_programs are expected to work in the elaborator.

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.