chore: Remove hir to ast pass (#5147)
# Description

## Problem\*

## Summary\*

Removes the Hir to Ast pass since it was only needed for the previously
proposed metacircular evaluator architecture where we'd globally loop
back to name resolution after the interpreter expands a macro. Now that
we're using an elaborator it is unneeded.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.