chore: Remove unused `new_variables` argument from `resolve_type_inner` (#5148)
# Description

## Problem\*

## Summary\*

This argument was previously used to add a fresh generic to a function's
signature when we found a slice type back when those used a generic for
the length. Since Slice types are now their own type this code path
isn't accessible anymore (*).

(\*) The codepath is technically accessible by trying to use a string
slice type but we don't support these yet anyway so I've removed them.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.