feat: cycle scalar <> bigfield interactions (#6744)
Adds missing methods in cycle_group and cycle_scalar required for the
PCS in the ECCVM recursive verifier. As a summary, we add missing
cycle_group constructors, a cycle group representation of `one`, a naive
method to convert bigfield to cycle_scalar (but that adds constraints)
and unify the `batch_mul` interface with the others present in our
codebase (i.e. swap points and scalars).

---------

Co-authored-by: codygunton <codygunton@gmail.com>