chore: start moving lints into a separate linting directory (#5165)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR starts moving some diagnostics into a separate `lints` directory
and out of the main flow of the elaborator.

Eventually we can ideally move this to run on the node interner after
the elaborator has completed rather than being called during elaboration
but this can happen in a later PR

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.