feat!: add session id to foreign call RPC requests (#5205)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds an `id` field to RPC foreign call requests so that the
TriXE is able to maintain separate internal state for different test
functions so that Aztec integration tests can be run in parallel. This
has necessitated some changes to the request format so this is a
breaking change.


The RPC request now looks like

```
'{
    "jsonrpc":"2.0",
    "method":"resolve_foreign_call",
    "params":[{
        "session_id":3789997497881369652,
        "function": "foo",
        "inputs": [
            "0000000000000000000000000000000000000000000000000000000000000001", 
["0000000000000000000000000000000000000000000000000000000000000001","0000000000000000000000000000000000000000000000000000000000000002"
    ]]
}]}
```

## Additional Context



## Documentation\*

Check one:
- [] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.