chore: move implementation of bitwise operations into `blackbox_solver` (#5209)
# Description

## Problem\*

Related to #5055 
## Summary\*

We don't need the field element to know about AND and XOR as we just
need to be able to convert into byte arrays and back. I've then moved
this into the blackbox solver crate (the only place we were actually
using this) to simplify the `AcirField` trait.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.