chore!: remove `param_witnesses` and `return_witnesses` from ABI (#5154)
# Description

## Problem\*

Followup to #5142 
Related to #5104 

## Summary\*

This PR deletes the `param_witnesses` and `return_witnesses` fields from
the ABI as the locations of all inputs/return values are defined
implicitly by the types themselves.

I've left this as draft as there's some changes which need to be made to
the JS code which will require some extra work (it needs to be made
aware of the sizes of various types.)

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.