chore: avoid `bn254_blackbox_solver` polluting feature flags (#5141)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Depending on `bn254_blackbox_solver` results in the `bn254` feature flag
being applied to the `acir` crate. This is unnecessary as we can just
have this crate use a `GenericFieldElement` and construct it as a BN254
field.

This simplifies the lives of people attempting to use Noir with an
alternative field choice.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.